+++ DTCD-StyleSystem/.storybook/preview.js
@@ -8,4 +8,17 @@ export const parameters = {
       date: /Date$/,
     },
   },
+  backgrounds: {
+    default: 'white',
+    values: [
+      {
+        name: 'white',
+        value: '#fefefe',
+      },
+      {
+        name: 'black',
+        value: '#353542',
+      },
+    ],
+  },
 }