+++ DTCD-StyleSystem/src/base-components/components.js
@@ -21,6 +21,7 @@ import BaseExpanderGroup from './BaseExpanderGroup/BaseExpanderGroup';
 import BaseHeading from './BaseHeading/BaseHeading';
 import BaseFileLoader from './BaseFileLoader/BaseFileLoader';
 import BaseDropdown from './BaseDropdown/BaseDropdown';
+import BaseDateTimePicker from './BaseDateTimePicker/BaseDateTimePicker';
 
 export default [
   { name: 'base-button', baseClass: BaseButton },
@@ -46,4 +47,5 @@ export default [
   { name: 'datasource-select', baseClass: DataSourceSelect },
   { name: 'gauge-segment-builder', baseClass: GaugeSegmentBuilder },
   { name: 'base-dropdown', baseClass: BaseDropdown },
+  { name: 'date-time-picker', baseClass: BaseDateTimePicker },
 ];